[#62] feat : 브랜드 모달 관련 API 연동 중
- 인기 브랜드 조회 API 연동 완료
- 최근 선택한 브랜드 조회 API 테스트 필요
- 브랜드 검색 API 일시중지(서버와의 논의 필요)
