+++ src/config/queryKeys.ts
@@ -9,4 +9,6 @@
 
 export const queryKeys = {
   itemCategory: ['itemCategory'] as const,
+  brandRecentSelected: ['brandRecentSelected'] as const,
+  brandTop: ['brandTop'] as const,
 }
