+++ src/index.tsx
@@ -9,7 +9,14 @@ import { Global } from '@emotion/react'
 import App from './App'
 import { reset } from './components/styles'
 
-const queryClient = new QueryClient()
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: 0,
+      staleTime: 1000 * 60 * 60,
+    },
+  },
+})
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
